+++ .github/workflows/run-lmcpgen.yml
@@ -1,0 +1,58 @@
+on:
+    push:
+        paths:
+            - 'infrastructure/uxas/**'
+            - 'infrastructure/paths.sh'
+            - 'infrastructure/specs/**'
+            - 'resources/run-lmcpgen'
+            - 'resources/run_lmcpgen.py'
+    pull_request:
+        paths:
+            - 'infrastructure/uxas/**'
+            - 'infrastructure/paths.sh'
+            - 'infrastructure/specs/**'
+            - 'resources/run-lmcpgen'
+            - 'resources/run_lmcpgen.py'
+    workflow_dispatch:
+    schedule:
+        - cron: '20 3 * * 4'
+
+name: Test Running LmcpGen
+jobs:
+    build:
+        strategy:
+            fail-fast: false
+            matrix:
+                os: [ubuntu-18.04, ubuntu-20.04]
+                python-version: [3.7, 3.8]
+        runs-on: ${{ matrix.os }}
+        steps:
+            - uses: actions/checkout@v2
+              with:
+                  path: OpenUxAS
+
+            - name: Set up python ${{ matrix.python-version }}
+              uses: actions/setup-python@v2
+              with:
+                  python-version: ${{ matrix.python-version }}
+
+            - name: Set up java
+              uses: actions/setup-java@v1
+              with:
+                  java-version: '11'
+
+            - name: Install infrastructure
+              run: |
+                  cd OpenUxAS
+                  infrastructure/install -vv --no-gnat --no-java -y
+
+            - name: LMCP and Amase devel setup
+              run: |
+                  cd OpenUxAS
+                  ./anod devel-setup lmcp
+                  ./anod devel-setup amase
+
+            - name: Run LmcpGen
+              run: |
+                  cd OpenUxAS
+                  resources/run-lmcpgen
