+++ infrastructure/uxas/src/uxas/util/__init__.py
@@ -1,0 +1,3 @@
+"""UxAS until module."""
+
+from __future__ import annotations
